// apps/git-client/src/generated/RepositoryChangedEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RepositoryId } from "./RepositoryId";
import type { RepositoryInvalidation } from "./RepositoryInvalidation";

export type RepositoryChangedEvent = { repositoryId: RepositoryId, invalidations: Array<RepositoryInvalidation>, };